move testhelper into SDK, improve linting
