+++ .golangci.yml
@@ -96,4 +96,4 @@ issues:
     - 'cyclomatic complexity [0-9]+ of func `\(\*provider\)\.getConfig` is high'
     - "SA1019: s.server.IPv6 is deprecated"
   exclude-dirs:
-    - pkg/machines
+    - apis/machines
