+++ sdk/.golangci.yml
@@ -1,0 +1,39 @@
+# Copyright 2022 The Kubermatic Kubernetes Platform contributors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+#
+# This file contains *additional* linting rules that just apply to the SDK.
+# When running `make lint`, the SDK is linted twice, once with the repository
+# root's .golangci.yml and once with the SDK's config file.
+#
+
+run:
+  timeout: 10m
+  modules-download-mode: readonly
+
+linters:
+  enable:
+    - depguard
+  disable-all: true
+
+linters-settings:
+  depguard:
+    rules:
+      noreverse:
+        deny:
+          - { pkg: k8c.io/machine-controller/pkg, desc: SDK must not depend on the main module }
+
+issues:
+  exclude-files:
+    - zz_generated.*.go
